Localtime
This small change will calculate the snapshot start datetime according
to the local time of the server where this is run.  By default, all
snapshots are done in UTC, which means the daily snapshots that it
saves will be at midnight UTC.  WIth this change the daily snapshots
that are kept will be at midnight localtime.

This is necessary because the snap_age is calculated by taking the
Time.now function minus the snap startTime.  If the server this is run
from is set to a timezone other than UTC, then this will give a skewed
snap age.  This fix accounts for that.